Tocify custom field instead of Content
Although the "Content" field is the main candidate for
silverstripe-autotoc, do not assume it. Allow instead to customize the
field name on a class basis by setting the MyClass.content_field
property via YAML, e.g.:

ProductPage_Controller:
    content_field: 'Details'
AuthorHandler:
    content_field: 'Biography'

Keep in mind Autotoc must be applied to a *controller*.

If not specified, 'Content' is assumed so backward compatibility is
retained.